+++ start/src/App.vue
@@ -6,6 +6,7 @@
   <dynamic-island-vue />
   <f-alert>123</f-alert>
   <f-button>123</f-button>
+  <f-empty image-src="https://tianyuhao.cn/images/auto/1.jpg" image-size="200px" />
 </template>
 
 <style lang="scss" scoped></style>